// frontend/StockManager/generated-sources/openapi/models/PortfolioItemDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * StockManager API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { StockDto } from './StockDto';
import {
    StockDtoFromJSON,
    StockDtoFromJSONTyped,
    StockDtoToJSON,
    StockDtoToJSONTyped,
} from './StockDto';

/**
 * 
 * @export
 * @interface PortfolioItemDto
 */
export interface PortfolioItemDto {
    /**
     * 
     * @type {number}
     * @memberof PortfolioItemDto
     */
    id?: number;
    /**
     * 
     * @type {StockDto}
     * @memberof PortfolioItemDto
     */
    stock?: StockDto;
    /**
     * 
     * @type {number}
     * @memberof PortfolioItemDto
     */
    quantity?: number;
    /**
     * 
     * @type {number}
     * @memberof PortfolioItemDto
     */
    purchasePrice?: number;
    /**
     * 
     * @type {number}
     * @memberof PortfolioItemDto
     */
    targetPrice?: number;
    /**
     * 
     * @type {string}
     * @memberof PortfolioItemDto
     */
    note?: string | null;
    /**
     * 
     * @type {number}
     * @memberof PortfolioItemDto
     */
    profitLoss?: number;
}

/**
 * Check if a given object implements the PortfolioItemDto interface.
 */
export function instanceOfPortfolioItemDto(value: object): value is PortfolioItemDto {
    return true;
}

export function PortfolioItemDtoFromJSON(json: any): PortfolioItemDto {
    return PortfolioItemDtoFromJSONTyped(json, false);
}

export function PortfolioItemDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): PortfolioItemDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'stock': json['stock'] == null ? undefined : StockDtoFromJSON(json['stock']),
        'quantity': json['quantity'] == null ? undefined : json['quantity'],
        'purchasePrice': json['purchasePrice'] == null ? undefined : json['purchasePrice'],
        'targetPrice': json['targetPrice'] == null ? undefined : json['targetPrice'],
        'note': json['note'] == null ? undefined : json['note'],
        'profitLoss': json['profitLoss'] == null ? undefined : json['profitLoss'],
    };
}

export function PortfolioItemDtoToJSON(json: any): PortfolioItemDto {
    return PortfolioItemDtoToJSONTyped(json, false);
}

export function PortfolioItemDtoToJSONTyped(value?: PortfolioItemDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'stock': StockDtoToJSON(value['stock']),
        'quantity': value['quantity'],
        'purchasePrice': value['purchasePrice'],
        'targetPrice': value['targetPrice'],
        'note': value['note'],
        'profitLoss': value['profitLoss'],
    };
}